#8 - resolved conflicts when refactoring code
