+++ src/main/java/com/example/splitly/application/serviceInterface/IUserInfo.java
@@ -1,0 +1,6 @@
+package com.example.splitly.application.serviceInterface;
+
+
+public interface IUserInfo {
+
+}
