+++ src/main/java/com/example/splitly/presentation/controller/GroupUserController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.example.splitly.application.serviceInterface.IGroupUser;
 import com.example.splitly.presentation.dto.response.ResponseData;
-import com.example.splitly.service.IGroupUser;
 
 import lombok.RequiredArgsConstructor;
