--- src/main/java/com/example/splitly/service/InterfaceUserInfo.java
@@ -1,6 +1,0 @@
-package com.example.splitly.service;
-
-
-public interface InterfaceUserInfo {
-
-}
